fix: QOpProblem joint trait and added sigmoid
