+++ core/src/ops/nn/mod.rs
@@ -8,9 +8,12 @@ pub use self::softmax::{Softmax, SoftmaxExp};
 
 pub use crate::internal::*;
 
+use tract_num_traits::AsPrimitive;
+
 element_wise!(sigmoid, Sigmoid,
  [f16] => |_, xs| { (tract_linalg::ops().sigmoid_f16)().run(xs) },
  [f32] => |_, xs| { (tract_linalg::ops().sigmoid_f32)().run(xs) };
+ q: [i8, u8, i32, i32] => |x: f32| 1.0 / (1.0+(-x).exp());
  cost: |dt| {tvec!((Cost::FMA(dt), 11), (Cost::Div(dt), 1))}
 );
