+++ test-rt/suite-unit/src/q_helpers.rs
@@ -52,3 +52,38 @@ pub fn qu8_tensor0(value: u8, zp: i32, scale: f32) -> TractResult<Tensor> {
 pub fn qu8_tensor1(values: &[u8], zp: i32, scale: f32) -> TractResult<Tensor> {
     qu8_tensor(tensor1(values), zp, scale)
 }
+
+pub trait QOpProblem {
+    fn reference_float_ops(&self) -> TractResult<Tensor>;
+
+    fn check_ref_with_approx(&self, result: Tensor, approx: Approximation) -> infra::TestResult {
+        let mut reference = self.reference_float_ops()?;
+        let out_dt = result.datum_type();
+        let (zero_point, scale) = out_dt.zp_scale();
+        let min_repr_val =
+            (out_dt.unquantized().min_value().cast_to_scalar::<f32>()? - zero_point as f32) * scale;
+        let max_repr_val =
+            (out_dt.unquantized().max_value().cast_to_scalar::<f32>()? - zero_point as f32) * scale;
+
+        reference
+            .to_array_view_mut()?
+            .iter_mut()
+            .for_each(|x: &mut f32| *x = (*x).clamp(min_repr_val, max_repr_val));
+
+        let mut fp_results = result.cast_to::<f32>()?.into_owned();
+
+        let acceptable_scale_error_ratio = match approx {
+            Approximation::Exact => 0.,
+            Approximation::Approximate => 2.,
+            _ => 3.,
+        };
+        assert!(tract_core::ndarray::Zip::from(fp_results.to_array_view_mut()?)
+            .and(reference.to_array_view()?)
+            .all(|x: &mut f32, xref: &f32| {
+                let closest_x = (*x).clamp(min_repr_val, max_repr_val);
+                // core maximal accepted distance by default
+                (xref - closest_x).abs() <= scale * acceptable_scale_error_ratio
+            }));
+        Ok(())
+    }
+}
